timers: Split out timer_types.h
commit dff0fd233a5104337069603d201f8cad74bc0e5a upstream.

Cutting down on sched.h dependencies: this is going to be used in
workqueue_types.h in the next patch, so we can kill the sched.h
dependency on workqueue.h.
